[SQL] Support LOCAL views, which do not produce circuit outputs (#1851)
