+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/parser/SqlLateness.java
@@ -1,13 +1,11 @@
-package org.dbsp.sqlCompiler.compiler.frontend.statements;
+package org.dbsp.sqlCompiler.compiler.frontend.parser;
 
 import org.apache.calcite.sql.SqlCall;
 import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.sql.SqlLiteral;
 import org.apache.calcite.sql.SqlNode;
-import org.apache.calcite.sql.SqlNodeList;
 import org.apache.calcite.sql.SqlOperator;
-import org.apache.calcite.sql.SqlSelect;
 import org.apache.calcite.sql.SqlSpecialOperator;
 import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.parser.SqlParserPos;
@@ -20,41 +18,38 @@ import org.checkerframework.dataflow.qual.Pure;
 import java.util.List;
 import java.util.Objects;
 
-/** This implements our own extension of SQL DML statement.
- * The REMOVE statement is almost like an INSERT statement,
- * but the effect is to remove a value from a table.
- * Unlike DELETE, the values are not specified by a query. */
-public class SqlRemove extends SqlCall {
+/** The LATENESS statement is used to declare the lateness of a view's columns. */
+public class SqlLateness extends SqlCall {
     public static final SqlSpecialOperator OPERATOR =
-            new SqlSpecialOperator("REMOVE", SqlKind.DELETE) {
+            new SqlSpecialOperator("LATENESS", SqlKind.OTHER) {
                 @SuppressWarnings("argument.type.incompatible")
                 @Override public SqlCall createCall(@Nullable SqlLiteral functionQualifier,
                                                     SqlParserPos pos,
                                                     @Nullable SqlNode... operands) {
-                    return new SqlRemove(
+                    return new SqlLateness(
                             pos,
-                            Objects.requireNonNull(operands[0]),
-                            Objects.requireNonNull(operands[1]),
-                            (SqlNodeList) operands[2]);
+                            (SqlIdentifier) Objects.requireNonNull(operands[0]),
+                            (SqlIdentifier) Objects.requireNonNull(operands[1]),
+                            Objects.requireNonNull(operands[2]));
                 }
             };
 
-    SqlNode targetTable;
-    SqlNode source;
-    @Nullable SqlNodeList columnList;
+    SqlIdentifier view;
+    SqlIdentifier column;
+    SqlNode expression;
 
-    public SqlRemove(SqlParserPos pos,
-                     SqlNode targetTable,
-                     SqlNode source,
-                     @Nullable SqlNodeList columnList) {
+    public SqlLateness(SqlParserPos pos,
+                       SqlIdentifier view,
+                       SqlIdentifier column,
+                       SqlNode expression) {
         super(pos);
-        this.targetTable = targetTable;
-        this.source = source;
-        this.columnList = columnList;
+        this.view = view;
+        this.column = column;
+        this.expression = expression;
     }
 
     @Override public SqlKind getKind() {
-        return SqlKind.DELETE;
+        return SqlKind.OTHER;
     }
 
     @Override public SqlOperator getOperator() {
@@ -63,61 +58,54 @@ public class SqlRemove extends SqlCall {
 
     @SuppressWarnings("nullness")
     @Override public List<SqlNode> getOperandList() {
-        return ImmutableNullableList.of(targetTable, source);
+        return ImmutableNullableList.of(view, column, expression);
     }
 
     @SuppressWarnings("assignment.type.incompatible")
     @Override public void setOperand(int i, @Nullable SqlNode operand) {
+        Objects.requireNonNull(operand);
         switch (i) {
             case 0:
                 assert operand instanceof SqlIdentifier;
-                this.targetTable = operand;
+                this.view = (SqlIdentifier) operand;
                 break;
             case 1:
-                this.source = Objects.requireNonNull(operand);
+                assert operand instanceof SqlIdentifier;
+                this.column = (SqlIdentifier) operand;
                 break;
             case 3:
-                this.columnList = (SqlNodeList) operand;
+                this.expression = operand;
                 break;
             default:
                 throw new AssertionError(i);
         }
     }
 
-    /** Return the identifier for the target table of the removal. */
-    public SqlNode getTargetTable() {
-        return targetTable;
+    public SqlIdentifier getView() {
+        return this.view;
     }
 
-    /** Returns the source expression for the data to be removed. */
-    public SqlNode getSource() {
-        return source;
-    }
-
-    public void setSource(SqlSelect source) {
-        this.source = source;
+    public SqlIdentifier getColumn() {
+        return this.column;
     }
 
     @Pure
-    public @Nullable SqlNodeList getTargetColumnList() {
-        return this.columnList;
+    public SqlNode getLateness() {
+        return this.expression;
     }
 
     @Override public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
         writer.startList(SqlWriter.FrameTypeEnum.SELECT);
-        writer.sep("REMOVE FROM");
+        writer.keyword("LATENESS");
         final int opLeft = getOperator().getLeftPrec();
         final int opRight = getOperator().getRightPrec();
-        this.targetTable.unparse(writer, opLeft, opRight);
-        if (this.columnList != null) {
-            this.columnList.unparse(writer, opLeft, opRight);
-        }
+        this.view.unparse(writer, opLeft, opRight);
+        this.column.unparse(writer, opLeft, opRight);
+        this.expression.unparse(writer, 0, 0);
         writer.newlineAndIndent();
-        this.source.unparse(writer, 0, 0);
     }
 
     @Override public void validate(SqlValidator validator, SqlValidatorScope scope) {
-        // validator.validateInsert(this);
         // TODO
     }
 }
