+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/compiler/frontend/parser/SqlRemove.java
@@ -1,11 +1,13 @@
-package org.dbsp.sqlCompiler.compiler.frontend.statements;
+package org.dbsp.sqlCompiler.compiler.frontend.parser;
 
 import org.apache.calcite.sql.SqlCall;
 import org.apache.calcite.sql.SqlIdentifier;
 import org.apache.calcite.sql.SqlKind;
 import org.apache.calcite.sql.SqlLiteral;
 import org.apache.calcite.sql.SqlNode;
+import org.apache.calcite.sql.SqlNodeList;
 import org.apache.calcite.sql.SqlOperator;
+import org.apache.calcite.sql.SqlSelect;
 import org.apache.calcite.sql.SqlSpecialOperator;
 import org.apache.calcite.sql.SqlWriter;
 import org.apache.calcite.sql.parser.SqlParserPos;
@@ -18,38 +20,41 @@ import org.checkerframework.dataflow.qual.Pure;
 import java.util.List;
 import java.util.Objects;
 
-/** The LATENESS statement is used to declare the lateness of a view's columns. */
-public class SqlLateness extends SqlCall {
+/** This implements our own extension of SQL DML statement.
+ * The REMOVE statement is almost like an INSERT statement,
+ * but the effect is to remove a value from a table.
+ * Unlike DELETE, the values are not specified by a query. */
+public class SqlRemove extends SqlCall {
     public static final SqlSpecialOperator OPERATOR =
-            new SqlSpecialOperator("LATENESS", SqlKind.OTHER) {
+            new SqlSpecialOperator("REMOVE", SqlKind.DELETE) {
                 @SuppressWarnings("argument.type.incompatible")
                 @Override public SqlCall createCall(@Nullable SqlLiteral functionQualifier,
                                                     SqlParserPos pos,
                                                     @Nullable SqlNode... operands) {
-                    return new SqlLateness(
+                    return new SqlRemove(
                             pos,
-                            (SqlIdentifier) Objects.requireNonNull(operands[0]),
-                            (SqlIdentifier) Objects.requireNonNull(operands[1]),
-                            Objects.requireNonNull(operands[2]));
+                            Objects.requireNonNull(operands[0]),
+                            Objects.requireNonNull(operands[1]),
+                            (SqlNodeList) operands[2]);
                 }
             };
 
-    SqlIdentifier view;
-    SqlIdentifier column;
-    SqlNode expression;
+    SqlNode targetTable;
+    SqlNode source;
+    @Nullable SqlNodeList columnList;
 
-    public SqlLateness(SqlParserPos pos,
-                       SqlIdentifier view,
-                       SqlIdentifier column,
-                       SqlNode expression) {
+    public SqlRemove(SqlParserPos pos,
+                     SqlNode targetTable,
+                     SqlNode source,
+                     @Nullable SqlNodeList columnList) {
         super(pos);
-        this.view = view;
-        this.column = column;
-        this.expression = expression;
+        this.targetTable = targetTable;
+        this.source = source;
+        this.columnList = columnList;
     }
 
     @Override public SqlKind getKind() {
-        return SqlKind.OTHER;
+        return SqlKind.DELETE;
     }
 
     @Override public SqlOperator getOperator() {
@@ -58,54 +63,61 @@ public class SqlLateness extends SqlCall {
 
     @SuppressWarnings("nullness")
     @Override public List<SqlNode> getOperandList() {
-        return ImmutableNullableList.of(view, column, expression);
+        return ImmutableNullableList.of(targetTable, source);
     }
 
     @SuppressWarnings("assignment.type.incompatible")
     @Override public void setOperand(int i, @Nullable SqlNode operand) {
-        Objects.requireNonNull(operand);
         switch (i) {
             case 0:
                 assert operand instanceof SqlIdentifier;
-                this.view = (SqlIdentifier) operand;
+                this.targetTable = operand;
                 break;
             case 1:
-                assert operand instanceof SqlIdentifier;
-                this.column = (SqlIdentifier) operand;
+                this.source = Objects.requireNonNull(operand);
                 break;
             case 3:
-                this.expression = operand;
+                this.columnList = (SqlNodeList) operand;
                 break;
             default:
                 throw new AssertionError(i);
         }
     }
 
-    public SqlIdentifier getView() {
-        return this.view;
+    /** Return the identifier for the target table of the removal. */
+    public SqlNode getTargetTable() {
+        return targetTable;
     }
 
-    public SqlIdentifier getColumn() {
-        return this.column;
+    /** Returns the source expression for the data to be removed. */
+    public SqlNode getSource() {
+        return source;
+    }
+
+    public void setSource(SqlSelect source) {
+        this.source = source;
     }
 
     @Pure
-    public SqlNode getLateness() {
-        return this.expression;
+    public @Nullable SqlNodeList getTargetColumnList() {
+        return this.columnList;
     }
 
     @Override public void unparse(SqlWriter writer, int leftPrec, int rightPrec) {
         writer.startList(SqlWriter.FrameTypeEnum.SELECT);
-        writer.keyword("LATENESS");
+        writer.sep("REMOVE FROM");
         final int opLeft = getOperator().getLeftPrec();
         final int opRight = getOperator().getRightPrec();
-        this.view.unparse(writer, opLeft, opRight);
-        this.column.unparse(writer, opLeft, opRight);
-        this.expression.unparse(writer, 0, 0);
+        this.targetTable.unparse(writer, opLeft, opRight);
+        if (this.columnList != null) {
+            this.columnList.unparse(writer, opLeft, opRight);
+        }
         writer.newlineAndIndent();
+        this.source.unparse(writer, 0, 0);
     }
 
     @Override public void validate(SqlValidator validator, SqlValidatorScope scope) {
+        // validator.validateInsert(this);
         // TODO
     }
 }
